requirements.txt
# local package
-e .

# external requirements
click~=8.1.3
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
kaggle~=1.5.12
pandas~=1.4.2
sklearn~=0.0
scikit-learn~=1.1.1
matplotlib~=3.5.2
seaborn~=0.11.2
pip~=22.1.2
tornado~=6.1
setuptools~=56.0.0
docutils~=0.16
numpy~=1.22.4